Life & Mana cannot be filled past their maximum values

+++ src/org/escaperun/game/model/entities/statistics/Level.java
@@ -14,8 +14,7 @@ public class Level extends DerivedStatistic<Integer> {
 
     @Override
     protected void recalculate() {
-        setBase(experience.getCurrent().intValue() / divisor);
-
+        setBase((int) experience.getCurrent().doubleValue() / divisor);
     }
 
     @Override
